Make endpoints restful
