+++ app/controllers/application_controller.rb
@@ -1,2 +1,7 @@
 class ApplicationController < ActionController::API
+  private
+
+  def encode_token(payload)
+    JWT.encode(payload, Rails.application.secrets.secret_key_base)
+  end
 end
